+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.3
+* Added border feature. Use borderWidth and borderColor to setup border.
+* Added rectangle shape. To use rectangle, specify shapeType: ShapeType.rectangle.
+* Added support for 2 letters.
+
 ## 0.0.2
 
 * Better letter M and W positioning